refactor(schedule): change file name from timetable to schedule
