+++ backend/app/main.py
@@ -1,7 +1,5 @@
 from fastapi import FastAPI
-from app.routers import test, rooms, programs, users, teacher, timetables, program_courses
-# Import models để register với SQLAlchemy
-import app.models
+from app.routers import test, rooms, programs, users, teacher, program_courses,schedules
 
 # Tạo app
 app = FastAPI(
@@ -15,5 +13,5 @@ app.include_router(rooms.router, prefix="/api")
 app.include_router(programs.router, prefix="/api")
 app.include_router(users.router, prefix="/api")
 app.include_router(teacher.router, prefix="/api")
-app.include_router(timetables.router, prefix="/api")
+app.include_router(schedules.router, prefix="/api")
 app.include_router(program_courses.router, prefix="/api")
